+++ tests/lobster/model/test__ume_onnx.py
@@ -6,7 +6,7 @@ from torch import Size, Tensor
 
 from lobster.constants import Modality
 from lobster.model import UME
-from lobster.model.onnx_utils import run_onnx_inference, benchmark_onnx_pytorch
+from lobster.model._onnx_utils import run_onnx_inference, benchmark_onnx_pytorch
 
 # Performance optimizations for CI/CD:
 # - Reduced max_length from 128 to 64
